+++ lib/src/texadd/noise_texture_adder.cpp
@@ -15,7 +15,7 @@ std::shared_ptr<Mesh> NoiseTextureAdder::generate(Mesh* before) {
   auto result = std::make_shared<Mesh>(*before);
 
   auto noiseModule = evaluateGraph(noiseGraph);
-  auto colorFunction = [&](Eigen::Vector3d worldPos) {
+  auto colorFunction = [&](Eigen::Vector3f worldPos) {
     if (noiseModule == nullptr) return coloring.colorFromValue(0);
     float value = (noiseModule->GetValue(worldPos.x(), worldPos.y(), worldPos.z()) + 1) / 2;
     return coloring.colorFromValue(value);
